+++ QuickSubtitleTranslator.IntegrationTestsMockEx/GeneralTests.cs
@@ -14,7 +14,8 @@ namespace QuickSubtitleTranslator.IntegrationTestsMockEx
     {
         public GeneralTests()
         {
-            File.WriteAllText("accepted_notice", "Yes");
+            if (!File.Exists("accepted_notice"))
+                File.WriteAllText("accepted_notice", "Yes");
         }
 
         [Fact]
@@ -23,27 +24,27 @@ namespace QuickSubtitleTranslator.IntegrationTestsMockEx
             /*
              * --path "..\..\..\..\test_folder" --output-folder "sub_output" --from-lang "en" --to-lang "es" --api "Google"
              */
-            string path = @"..\..\..\..\test_folder";
+            string path = @"..\..\..\..\mockex_test_OutputSubtitle_ShouldBe_Exact";
             string outputFolder = "sub_output";
             string fromLang = "en";
             string toLang = "es";
             string apiKey = "";
-            APIType api = APIType.Google;
+            ApiType api = ApiType.Google;
 
             if (Directory.Exists("sub_output"))
             {
                 File.Delete("sub_output\\Scrubs.S02E08.srt");
                 File.Delete("sub_output\\srt example.srt");
                 File.Delete("sub_output\\sub_example.srt");
+
+                File.Delete("sub_output\\Scrubs.S02E08_nf.srt");
+                File.Delete("sub_output\\srt example_nf.srt");
+                File.Delete("sub_output\\sub_example_nf.srt");
             }
 
-            //Dirty fix so it doesn't break up lines
-            Constants.DoNotBreakIfOnlyLine = 1000;
-            Constants.MaxWordsPerLine = 1000;
-
             var svc = new Mock<ITranslationService>();
-            svc.Setup(x => x.Translate(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<IReadOnlyList<MySubtitleItem>>(), It.IsAny<string>()))
-                .Returns((string to, string from, IReadOnlyList<MySubtitleItem> list, string key) =>
+            svc.Setup(x => x.Translate(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<IReadOnlyList<MySubtitleItem>>(), It.IsAny<string>(), It.IsAny<bool>()))
+                .Returns((string to, string from, IReadOnlyList<MySubtitleItem> list, string key, bool waitForInput) =>
                 {
                     var listx = new List<MyTranslatedSubtitleItem>(list.Count);
                     foreach (var item in list)
@@ -53,23 +54,16 @@ namespace QuickSubtitleTranslator.IntegrationTestsMockEx
                     return new MyTranslateResult(listx, 0);
                 });
 
-            Program.TranslationService = svc.Object;
-            Program.Main(path, outputFolder, fromLang, toLang, api, apiKey);
-
-            byte[] md5Source;
-            byte[] md5Output;
-            using (var md5 = MD5.Create())
+            var app = new App()
             {
-                using var stream = File.OpenRead(path + @"\" + "srt example.srt");
-                md5Source = md5.ComputeHash(stream);
-            }
-            using (var md5 = MD5.Create())
-            {
-                using var stream = File.OpenRead(@"sub_output\" + "srt example.srt");
-                md5Output = md5.ComputeHash(stream);
-            }
+                TranslationService = svc.Object,
+                //Dirty fix so it doesn't break up lines
+                DoNotBreakIfOnlyLine = 1000,
+                MaxWordsPerLine = 1000
+            };
+            app.Run(path, outputFolder, fromLang, toLang, api, apiKey);
 
-            Assert.True(md5Source.SequenceEqual(md5Output), "Files are not identical");
+            Helper.AssertFilesAreIdentical(path + @"\" + "srt example.srt", @"sub_output\" + "srt example.srt");
         }
     }
 }
